feat: Rename ArrayView #156
