--- frontend/src/pages/archive/tab/Group.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-type Props = {}
-
-const Group = (props: Props) => {
-  return (
-    <div>그룹별로 조회 후 일괄입력 체크</div>
-  )
-}
-
-export default Group
